// src/client/src/screens/common/detail.tsx
export const StyledOption = styled.div<{ $width: string, $height: string, $borderRadius: string }>`
    width: ${props => props.$width};
    height: ${props => props.$height};
    background-color: #fff;
    border-radius: ${props => props.$borderRadius};

    display: flex;
    justify-content: center;
    align-items: center;
`